fixed node icon displaying
